Add function for fixing web address server and exeption hadling
If the user inputs a web address with http or https or/and forward
slashes after the TLD domain name a function is called that clears
the web address with a regex.

If a user enters a web server address with support only for http
program will timeout after 5 seconds and print out an error message to
the user. Also if a non existant domain name is entered an execption is
caught and the program will exit with the correct error code.

Fix a mistake in docs for a function.