Improve port discovery
Implement a mechanism to combat protection of sending too many
requests in a time window. When a connection is atempted and if
an exception is thrown that signalizes blocking of requests becuase
too many were sent the program sleeps for 1 second incremeting by one
each time it sleeps. After it wakes up it sends another request.
After 15 seconds it gives up.

Rename some options for clarity.