Refractor certificate and cipher suite parameters
Booth classes extend an abstract class `Parameters` which defines
a method for rating parameters.

All of cipher suite parameters are in a nested dictionary in
parameters dictionary for easier convertion to json later on.

Non-ratable certificate parameters such as serial number are now in
a separate dictionary that just holds a list. Other ratable
parameters are stored in a nested dictionary same as cipher
suite parameters.