Adapt script to use the penterep mail template
Src folder containts all the source files for the script.

Moved the ssltest script to a separte package src so so that
relative imports work correctly inside the package src and all the
subpacakges.

Main script file is now the template file SSLTester.py which runs
start_scan.py file from main package.

This is a temporay state just to merge the two projects together.