Add SSLv2 main connection parameters and fix TLSv1.0 bug
Since the cipher suite is choosen by the client and not the server a
cipher suite is hardcoded into the `parse_cipher_suite` function since
there are only 7 cipher suites in SSLv2 and all of them are using MD5
hashing which is insecure for now it doesn't matter which one is
choosen.

If multiple cipher suite scanning is implemented the hardcoded cipher
suite can be changed.

Also fix the main connection displaying TLSv1 instead of TLSv1.0