Remake vulnerabilities scanning/testing, improve logging
Since it is more secure to scan on every valid SSL/TLS version for
vulnerabilities one type of scan needs to run one or more tests.

Vulnerabilities are inheriting from VulnerabilityTest class which
defines a scan method which runs tests for every valid SSL/TLS
version. These tests are run in parallel. If one test is positive
all of other tests are canceled.

This change allows for the improvement of logging for vulnerabilities.
New logging is implemented in VulnerabillityTest class so every newly
added vulnerability doesn't need to implement logging all over again.

Reduce the level of logging for SSL/TLS protocol version scanning to
info instead of debug.