Add more OpenSSL ciphers conversions, change filling of cipher suites
More OpenSSL cipher suites conversions are added to the
cipher_suites.json file.

Filling of cipher suites in ClientHello class is changed. Either a json
file with cipher suites in hex string format is read only for SSLv3.
For other protocols the ssl library is used to read the cipher suites.