Implement an improved and colorized text output for scans
Address (IP, port) and for each category (parameters, supported
protocols) a title is printed out with a special color. Titles are also
filled with special characters such as "=" or "-" so that the title
fills the terminal width.

Values for parmeters such as ratings for cipher suites are given color
coresponding to their meaming. For example 1 is green and 4 is red.